fix(style): fix style shortcode randomly not rendered issue
Possibly due to the new rendering order started from Hugo v0.123.0. The `styleArr` in style.html is set AFTER it is read in assets.html. Changing the key in Scratch (now aliased to Store) seems to magically fix the issue, so I'm not going to invest more time investigating it.